Add missing domain check to `isForumRoute` and `isCpRoute`
Currently, the `Waterhole::isForumRoute()` and `Waterhole::isCpRoute()` methods only check that the request path matches.

This creates issues when a custom domain is specified in `config/system.php`, especially when integrated into an existing Laravel project.

This commit adds a check to ensure the domain matches if one is specified in the `system` config.